Get user by fields complete and move helpers to own file
